Add state, factory and serializer for Int128 type The dedicated state, serializer and factory were added to replace LongDecimalWithOverflowAndLongState in avg and sum functions for decimals. LongDecimalWithOverflowAndLongState, LongDecimalWithOverflowState and theirs factories and seralizers were deleted. (before) BenchmarkDecimalAggregation.benchmark                             avg    LONG  avgt    6      8.584 ±   0.046  ns/op (after)  BenchmarkDecimalAggregation.benchmark                             avg    LONG  avgt    6      8.919 ±   0.040  ns/op (before) BenchmarkDecimalAggregation.benchmarkEvaluateFinal                avg    LONG  avgt    6  84641.569 ± 104.723  ns/op (after)  BenchmarkDecimalAggregation.benchmarkEvaluateFinal                avg    LONG  avgt    6  99071.053 ± 144.885  ns/op (before) BenchmarkDecimalAggregation.benchmarkEvaluateIntermediate         avg    LONG  avgt    6      8.604 ±   0.032  ns/op (after)  BenchmarkDecimalAggregation.benchmarkEvaluateIntermediate         avg    LONG  avgt    6      8.967 ±   0.070  ns/op
